Remove `api_version` from `Context.org`
Since it's confusing to have two API versions - the org API version
and the one that (in the future) will be from `pyproject.toml`.

We can always add this back later if it's really needed, however at
that point we will likely want to pick a name that makes the distinction
between the two versions clearer.