Add support for binary fields (#6)
Adds support for binary fields based on an allow-list of object types and field names. The allow-list is the same as the one used in the Java SDK implementation. Binary fields are supported for create, retrieve and update operations, tests have been added accordingly.

I chose to use the `bytes` type over `bytearray` for binary data fields as it is immutable and therefore easier to work with. However, when writing data, both types are supported to avoid that users have to transform their binary data which might be costly.

Closes: GUS-W-11790230