Add a CODEOWNERS file
For parity with our other repositories.

Also sets myself as the default reviewer for Dependabot PRs, to reduce
email notification noise to the rest of the members of the team alias, since
our teams have the "Only notify requested team members" option enabled.
(This is something we should roll out to our other repositories in the future.)

See:
https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#reviewers

GUS-W-12073227.