Switch `DataAPI` from stdlib's JSON to `orjson` too
The rest of the runtime uses `orjson` over the stdlib's `json` module,
since the former has much better performance, as previously confirmed
with benchmarks run against the runtime, and also as shown here:
https://github.com/ijl/orjson#performance

This switches `DataAPI` to using `orjson` too, for parity and performance.

As part of doing that, it has exposed that some of the type annotations for
`process_response` were incorrect - they were being passed an `Any` type,
that can include `None` (for example for delete requests that have a 204
response), but the existing type annotations didn't include `None`.

GUS-W-12098069.