Improve the CLI output for the `serve` and `check` commands
The `serve` command now outputs the version of the runtime being used,
as well as the configuration of the server (currently this is only the
number of workers, since other than host/port which are already output
by uvicorn, the number of workers is the only other config option).

The `check` command's output has been made more concise, since the
"Checking function..." intro was unnecessary (given the command runs
instantly).

GUS-W-12094313.
GUS-W-12206626.