Enable checking of code coverage
Locally, code coverage checking is enabled, but only prints a summary,
and does not enforce a threshold, since the common case of testing
locally is with the default pytest options, which intentionally do not
run the WireMock-requiring tests.

In CI however, code coverage is enforced, and is checked by running
pytest twice (once for tests that run on all platforms, and again for the
WireMock tests), followed by the `coverage report` command.

The `pytest-cov` pytest plugin is used, which is a thin wrapper around
the `coverage` Python package.

In the future additional tests will be added to cover some more of the
cases currently marked as `# pragma: no cover`.

See:
https://pytest-cov.readthedocs.io
https://coverage.readthedocs.io

GUS-W-12207644.