Test that the version command output can be parsed as semver
The output of the `sf-functions-python version` command is consumed
by the CLI (and in the future by the CNB) and parsed as semver, in order
to perform the minimum version check.

This test ensure that the output remains parsable like this in the future,
and that some other prefix isn't added to the version stdout, or a non-semver
version string (like `X.Y`) isn't used.

GUS-W-12207642.