Add README badges for PyPI and CI (#30)
To make the currently published version more discoverable, and to save
a few clicks when I (or anyone else) wants to visit PyPI or CI results.

I was going to also add the shields.io badge for listing the supported
Python versions, however that appears to be dependant on the PyPI
package `classifiers`, rather than using the newer `requires-python`
field. We'll add the classifiers later, but until they are set in the published
package, the badge would be broken, so I've not added it here.