Improve cleanup of the custom `sys.tracebacklimit` in tests
Previously the cleanup for a test was not run if the test failed prior
to the cleanup step, meaning that later unrelated tests (that depend
on traceback output matching a certain string) would fail, causing confusion
when debugging.

Now cleanup is run from a `finally` block, to ensure it always occurs.

GUS-W-12256166.