Refactor the internal CLI/ASGI app API for project path
It's not possible to directly pass options to an ASGI app run with unvicorn, so
the CLI relies upon setting an environment variable to pass the user-provided
path to the project directory to the ASGI app.

Until now, the utility function for this lived in the `config` module, however:
- this handling is not really configuration, but an internal implementation
  detail between the CLI and ASGI app
- we're about to add actual configuration (in the form of `project.toml` support),
  and if we put both in a `config` module it's going to be confusing.

As such, it has been moved into the `app` module, and inlined into the one place
where it is used.

The test has been removed, since the functionality is already tested by the app tests.

GUS-W-12256363.