Add a test for DataAPI's shared session handling
There were previously no tests in `test_data_api.py` that used a shared session,
even though a shared session is always used when the data API is initialised by
the ASGI app.

Whilst there is a test in `test_app.py` that tests the integration of the data API
into the ASGI app (and thus using a shared session), it only tests:
- the happy path (no request exceptions)
- making a single request (so doesn't test session re-use)

The new test tests both of these, allowing the removal of a few more coverage
exclusion annotations.

GUS-W-12261661.