Validate type annotations using pyright too
We already validate the Python type annotations using mypy, however
VSCode's Python integration uses the more-strict alternative type-checker
pyright.

This has meant on a few occasions type failures only show up in my IDE
and not in CI. (As demonstrated by the types that needed fixing as part
of this PR.)

To prevent this, we now run pyright in CI. It is run in both the standard
full-codebase type checking mode, as well as a mode that verifies the
type-completeness of our publicly exported types (which checks the
types from a consumer's point of view).

We still run mypy, since many consumers of this package may still be using
mypy, so we need to ensure our exported type annotations keep both type
checkers satisfied.

https://github.com/microsoft/pyright#installation
https://github.com/microsoft/pyright/blob/main/docs/configuration.md
https://github.com/microsoft/pyright/blob/main/docs/typed-libraries.md#verifying-type-completeness

GUS-W-12291199.