Make all module-internal functions private
Whilst all of these functions are descendants of the `_internal` module,
the documentation linters I was trying out do not recognise this, which
causes false positives.

Renaming these to be private both avoids the above, and makes it clearer
when looking at these modules that the functions are not only not publicly
exported, but are also only meant for usage within that single file.

GUS-W-12366695.