Update CloudEvent fixtures to more closely match production
There are example CloudEvents used in our unit tests and also the development
helper script `invoke.sh`, however their contents was incomplete and slightly
out of date.

I dumped the contents of a CloudEvent in production (by using a modified version
of the runtime package), and have used that to update the CloudEvent fixtures.

GUS-W-11879536.