Make `mock_event` generate a unique event ID each time it is called
Previously the ID was generated at import time and then would not change
from one invocation to the next.

Moving the `uuid4()` call out of the function signature also has the benefit
that the UUID string doesn't appear in the auto-generated docs, which was
causing the generated content to change every time the docs tool was run.

GUS-W-12442238.