Add docstrings to the data API class attributes missing them
The `SalesforceRestApiError`, `InnerSalesforceRestApiError` and `ReferenceId`
data API classes were missing docstrings on the class attributes, which results
in `pdoc3` auto-generating a confusing description, eg:

```
* `id: str`
    Return an attribute of instance, which is of type owner.
```

Now, the explicitly set docstrings are shown instead.

In addition, the order of the exceptions in `data_api.exceptions` has been adjusted,
so that the exceptions are displayed in the docs in a more intuitive order (base class
etc).

GUS-W-12442866.