Fix the content type and body handling of data API requests
Previously the values for `encoding` and `content-type` were inverted,
causing eg:

```
Traceback (most recent call last):
  File ".../.venv/lib/python3.10/site-packages/salesforce_functions/_internal/app.py", line 79, in _handle_function_invocation
    function_result = await function(event, context)
  File ".../functions/hellopython/main.py", line 77, in function
    result = await context.org.data_api.commit_unit_of_work(unit_of_work)
  File ".../.venv/lib/python3.10/site-packages/salesforce_functions/data_api/__init__.py", line 200, in commit_unit_of_work
    return await self._execute(
  File ".../.venv/lib/python3.10/site-packages/salesforce_functions/data_api/__init__.py", line 243, in _execute
    return await rest_api_request.process_response(response.status, json_body)
  File ".../.venv/lib/python3.10/site-packages/salesforce_functions/data_api/_requests.py", line 199, in process_response
    raise SalesforceRestApiError(
salesforce_functions.data_api.exceptions.SalesforceRestApiError: Salesforce REST API reported the following error(s):
---
UNSUPPORTED_MEDIA_TYPE error:
MediaType of 'utf-8' is not supported by this resource
```

In addition (but not the cause of the above error), for data API requests where
the body should be none (such as `query` and `delete`), previously the body
was still serialised to JSON, resulting in a body of `b'null'`. Now the body is empty
in those cases, and the `Content-Type` not set at all.

The Wiremock mappings have been updated to enforce that the `Content-Type` is
present and correct for requests that have a body, and that both the `Content-Type`
and body are absent when there shouldn't be a body. See:
https://wiremock.org/docs/request-matching/

GUS-W-12496700.