Make `mock_event` generate a unique event time each time it is called
Previously the time was generated at import time and then would not change
from one invocation to the next.

Moving the `datetime.today()` call out of the function signature also has the
benefit that the datetime string doesn't appear in the auto-generated docs, which
was causing the generated content to change every time the docs tool was run.

This same fix was applied to the `id` parameter, and should have been applied
to `time` at that point too.

GUS-W-12442238.